Renaming BeInListRequirement to CurrentUserInListRequirement as per review feedback.
